1.some fix 2.problem is about updating waiting time

+++ Include/ElevatorController.h
@@ -41,7 +41,7 @@ protected:
 	int aveWaitingTimeCount;
 	double aveWaitingTime;
 	double aveFlow;
-	int totalFlow;
+	double totalFlow;
 	/////////////////////////
 	/////////////////////////
 };
